Moved all obsoleted methods and properties to a single file and made partial classes where needed.
