+++ Assets/Mirror/Runtime/NetworkConnection.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 namespace Mirror
 {
-    public class NetworkConnection : IDisposable
+    public partial class NetworkConnection : IDisposable
     {
         public HashSet<NetworkIdentity> visList = new HashSet<NetworkIdentity>();
 
@@ -17,16 +17,6 @@ namespace Mirror
         public NetworkIdentity playerController { get; internal set; }
         public HashSet<uint> clientOwnedObjects;
         public bool logNetworkMessages;
-
-        // this is always true for regular connections, false for local
-        // connections because it's set in the constructor and never reset.
-        [Obsolete("isConnected will be removed because it's pointless. A NetworkConnection is always connected.")]
-        public bool isConnected { get; protected set; }
-
-        // this is always 0 for regular connections, -1 for local
-        // connections because it's set in the constructor and never reset.
-        [Obsolete("hostId will be removed because it's not needed ever since we removed LLAPI as default. It's always 0 for regular connections and -1 for local connections. Use connection.GetType() == typeof(NetworkConnection) to check if it's a regular or local connection.")]
-        public int hostId = -1;
 
         public NetworkConnection(string networkAddress)
         {
@@ -113,14 +103,6 @@ namespace Mirror
         public void UnregisterHandler(short msgType)
         {
             m_MessageHandlers.Remove(msgType);
-        }
-
-        [Obsolete("use Send<T> instead")]
-        public virtual bool Send(int msgType, MessageBase msg, int channelId = Channels.DefaultReliable)
-        {
-            // pack message and send
-            byte[] message = MessagePacker.PackMessage(msgType, msg);
-            return SendBytes(message, channelId);
         }
 
         public virtual bool Send<T>(T msg, int channelId = Channels.DefaultReliable) where T: IMessageBase
