+++ Assets/Mirror/Runtime/Obsolete.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 47440b8e8725b8543920eed6e4d01b1a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
